refactor(tangle-cloud): replace approval modals with unified service request detail modal
Refactor service request approval flow:
- Replace separate ApproveConfirmationModal and RejectConfirmationModal with unified ServiceRequestDetailModal
- Add ApprovedInstanceTable to show requests operator has approved/rejected
- Update PendingInstanceTable to filter out requests operator has already acted on
- Add "Review" button instead of separate Approve/Reject buttons
- Add new "Approved" tab in instances page for operator's acted requests
