+++ apps/tangle-cloud/src/pages/instances/Instances/index.tsx
@@ -1,5 +1,9 @@
 import { TableAndChartTabs } from '@tangle-network/ui-components/components/TableAndChartTabs';
-import { PlayFillIcon, TimeLineIcon } from '@tangle-network/icons';
+import {
+  PlayFillIcon,
+  TimeLineIcon,
+  CheckboxCircleLine,
+} from '@tangle-network/icons';
 import {
   FC,
   ReactElement,
@@ -11,17 +15,20 @@ import {
 import { TabContent } from '@tangle-network/ui-components';
 import { RunningInstanceTable } from './RunningInstanceTable';
 import { PendingInstanceTable } from './PendingInstanceTable';
+import { ApprovedInstanceTable } from './ApprovedInstanceTable';
 import useEvmOperatorInfo from '../../../hooks/useEvmOperatorInfo';
 import useOperatorStats from '../../../data/operators/useOperatorStats';
 
 enum InstancesTab {
   RUNNING_INSTANCES = 'Running',
   PENDING_INSTANCES = 'Pending',
+  APPROVED_INSTANCES = 'Approved',
 }
 
 const InstancesTabIcon: ReactElement[] = [
   <PlayFillIcon viewBox="0 0 16 16" className="w-4 h-4 !fill-blue-50" />,
   <TimeLineIcon className="w-4 h-4 !fill-yellow-100" />,
+  <CheckboxCircleLine className="w-4 h-4 !fill-green-50" />,
 ] as const;
 
 interface InstancesTabsProps {
@@ -44,13 +51,13 @@ export const InstancesTabs: FC<InstancesTabsProps> = ({
     return operatorStatsData && operatorStatsData.registeredBlueprints > 0;
   }, [operatorStatsData]);
 
-  const shouldShowPendingTab = isOperator && hasRegisteredBlueprints;
+  const shouldShowOperatorTabs = isOperator && hasRegisteredBlueprints;
 
-  const availableTabs = shouldShowPendingTab
+  const availableTabs = shouldShowOperatorTabs
     ? Object.values(InstancesTab)
     : [InstancesTab.RUNNING_INSTANCES];
 
-  const availableIcons = shouldShowPendingTab
+  const availableIcons = shouldShowOperatorTabs
     ? InstancesTabIcon
     : [InstancesTabIcon[0]];
 
@@ -77,12 +84,24 @@ export const InstancesTabs: FC<InstancesTabsProps> = ({
         />
       </TabContent>
 
-      {shouldShowPendingTab && (
+      {shouldShowOperatorTabs && (
         <TabContent
           value={InstancesTab.PENDING_INSTANCES}
           className="flex justify-center mx-auto"
         >
           <PendingInstanceTable
+            refreshTrigger={refreshTrigger}
+            setRefreshTrigger={setRefreshTrigger}
+          />
+        </TabContent>
+      )}
+
+      {shouldShowOperatorTabs && (
+        <TabContent
+          value={InstancesTab.APPROVED_INSTANCES}
+          className="flex justify-center mx-auto"
+        >
+          <ApprovedInstanceTable
             refreshTrigger={refreshTrigger}
             setRefreshTrigger={setRefreshTrigger}
           />
